+++ samples/cba_fused_infer/CMakeLists.txt
