+++ samples/utils.hpp
